+++ alacritty_terminal/src/event.rs
@@ -1,17 +1,12 @@
 use std::borrow::Cow;
 use std::fmt::{self, Debug, Formatter};
-use std::path::PathBuf;
 use std::sync::Arc;
 
-use crate::message_bar::Message;
 use crate::term::{ClipboardType, SizeInfo};
 
 #[derive(Clone)]
 pub enum Event {
-    DPRChanged(f64, (u32, u32)),
-    ConfigReload(PathBuf),
     MouseCursorDirty,
-    Message(Message),
     Title(String),
     ClipboardStore(ClipboardType, String),
     ClipboardLoad(ClipboardType, Arc<dyn Fn(&str) -> String + Sync + Send + 'static>),
@@ -23,10 +18,7 @@ pub enum Event {
 impl Debug for Event {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         match self {
-            Event::DPRChanged(scale, size) => write!(f, "DPRChanged({}, {:?})", scale, size),
-            Event::ConfigReload(path) => write!(f, "ConfigReload({:?})", path),
             Event::MouseCursorDirty => write!(f, "MouseCursorDirty"),
-            Event::Message(msg) => write!(f, "Message({:?})", msg),
             Event::Title(title) => write!(f, "Title({})", title),
             Event::ClipboardStore(ty, text) => write!(f, "ClipboardStore({:?}, {})", ty, text),
             Event::ClipboardLoad(ty, _) => write!(f, "ClipboardLoad({:?})", ty),