+++ alacritty_terminal/src/util.rs
@@ -1,6 +1,6 @@
 use std::ffi::OsStr;
+use std::io;
 use std::process::{Command, Stdio};
-use std::{cmp, io};
 
 #[cfg(not(windows))]
 use std::os::unix::process::CommandExt;
@@ -13,20 +13,16 @@ use winapi::um::winbase::{CREATE_NEW_PROCESS_GROUP, CREATE_NO_WINDOW};
 /// Threading utilities.
 pub mod thread {
     /// Like `thread::spawn`, but with a `name` argument.
-    pub fn spawn_named<F, T, S>(name: S, f: F) -> ::std::thread::JoinHandle<T>
+    pub fn spawn_named<F, T, S>(name: S, f: F) -> std::thread::JoinHandle<T>
     where
         F: FnOnce() -> T + Send + 'static,
         T: Send + 'static,
         S: Into<String>,
     {
-        ::std::thread::Builder::new().name(name.into()).spawn(f).expect("thread spawn works")
+        std::thread::Builder::new().name(name.into()).spawn(f).expect("thread spawn works")
     }
 
     pub use std::thread::*;
-}
-
-pub fn limit<T: Ord>(value: T, min: T, max: T) -> T {
-    cmp::min(cmp::max(value, min), max)
 }
 
 #[cfg(not(windows))]
@@ -78,16 +74,4 @@ where
         .creation_flags(CREATE_NEW_PROCESS_GROUP | CREATE_NO_WINDOW)
         .spawn()
         .map(|_| ())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::limit;
-
-    #[test]
-    fn limit_works() {
-        assert_eq!(10, limit(10, 0, 100));
-        assert_eq!(10, limit(5, 10, 100));
-        assert_eq!(100, limit(1000, 10, 100));
-    }
 }